Added a heuristic approach to saving context efficiently
This commit is a big one. We have reorganized the files that being
serve into seperated components. We now have JS and CSS file, hurray!
The main thing is that I've attempted to implement a heuristic approach
to shard the chat context into smaller, summarized pieces that can be
used so that we don't send a long context to the API at every call.